rename encoding to encode
