+++ docs/source/encode.md
@@ -1,0 +1,8 @@
+# Fermion-Qubit Encoding
+
+```{eval-rst}
+.. automodule:: ferrmion.encode
+   :members:
+   :undoc-members:
+   :show-inheritance:
+```
