--- docs/source/encoding.md
@@ -1,35 +1,0 @@
-# Base Encoding
-
-```{eval-rst}
-.. automodule:: ferrmion.encoding.base
-   :members:
-   :undoc-members:
-   :show-inheritance:
-```
-
-# Ternary Tree Encoding
-
-```{eval-rst}
-.. automodule:: ferrmion.encoding.ternary_tree
-   :members:
-   :undoc-members:
-   :show-inheritance:
-```
-
-# Ternary Tree Node
-
-```{eval-rst}
-.. automodule:: ferrmion.encoding.ternary_tree_node
-   :members:
-   :undoc-members:
-   :show-inheritance:
-```
-
-# k-NTO encoding
-
-```{eval-rst}
-.. automodule:: ferrmion.knto
-   :members:
-   :undoc-members:
-   :show-inheritance:
-```
