+++ docs/source/index.md
@@ -7,9 +7,9 @@
 
 ```{toctree}
 :caption: 'Contents:'
-:maxdepth: 2
+:maxdepth: 1
 
-encoding
+encode
 optimize
 devices
 utils
