+++ docs/source/optimize.md
@@ -1,7 +1,18 @@
-# Optimize Module
+# Optimize
+
+## Enumeration
 
 ```{eval-rst}
-.. automodule:: ferrmion.optimize
+.. automodule:: ferrmion.optimize.enumeration
+   :members:
+   :undoc-members:
+   :show-inheritance:
+```
+
+## Ternary Tree
+
+```{eval-rst}
+.. automodule:: ferrmion.optimize.rett
    :members:
    :undoc-members:
    :show-inheritance:
