+++ python/ferrmion/encode/ternary_tree.py
