+++ python/ferrmion/encode/ternary_tree_node.py
